# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SOAPP Documentación",
        "description": "Salud Ocupacional APP",
        "contact": {
            "email": "daruiza@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8080/api",
            "description": "SOAPP API Server"
        }
    ],
    "paths": {
...
        "/commerce/index": {
            "get": {
                "tags": [
                    "Commerce"
                ],
                "summary": "Get All Commerces",
                "description": "Return Commerces",
                "operationId": "getAllCommerces",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Commerce Name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nit",
                        "in": "query",
                        "description": "Commerce nit",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "department",
                        "in": "query",
                        "description": "Commerce department",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Commerce city",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Commerce User",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/commerce/store": {
            "post": {
                "tags": [
                    "Commerce"
                ],
                "summary": "Store Commerce",
                "description": "Store Commerce",
                "operationId": "storeCommerce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Commerce"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/commerce/showbyuserid/{id}": {
            "get": {
                "tags": [
                    "Commerce"
                ],
                "summary": "Get One Commerce",
                "description": "Return Commerce",
                "operationId": "get Commerce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/commerce/showbycommerceid/{id}": {
            "get": {
                "tags": [
                    "Commerce"
                ],
                "summary": "Get One Commerce By one Id",
                "description": "Return One Commerce",
                "operationId": "getCommerceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Commerce Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/commerce/update/{id}": {
            "put": {
                "tags": [
                    "Commerce"
                ],
                "summary": "Update One Commerce By one Id",
                "description": "Update One Commerce",
                "operationId": "getUpdateCommerceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Commerce Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Commerce"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/commerce/destroy/{id}": {
            "delete": {
                "tags": [
                    "Commerce"
                ],
                "summary": "Delete One Commerce By one Id",
                "description": "Delete One Commerce",
                "operationId": "getDestroyCommerceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Commerce Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Commerce": {
                "title": "Commerce",
                "description": "Commerce body data",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the commerce",
                        "type": "string",
                        "example": "TempoSolutions"
                    },
                    "nit": {
                        "title": "nit",
                        "description": "Nit of the Commerce",
                        "type": "string",
                        "example": "1039420535-3"
                    },
                    "department": {
                        "title": "department",
                        "description": "Department of the commerce",
                        "type": "string",
                        "example": "Antioquia"
                    },
                    "city": {
                        "title": "city",
                        "description": "Name of the commerce",
                        "type": "string",
                        "example": "Medellin"
                    },
                    "adress": {
                        "title": "adress",
                        "description": "adress of the commerce",
                        "type": "string",
                        "example": "Cll 60 # 75 - 150"
                    },
                    "description": {
                        "title": "description",
                        "description": "description of the commerce",
                        "type": "string",
                        "example": "super"
                    },
                    "logo": {
                        "title": "logo",
                        "description": "Logo of the commerce",
                        "type": "string",
                        "example": "default.png"
                    },
                    "user_id": {
                        "title": "user_id",
                        "description": "user_id of the commerce",
                        "type": "string",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}